fix: altera dto de retorno dos insights
