+++ src/main/java/com/vulpix/api/Dto/Publicacao/Insights/PublicacaoInsightDto.java
@@ -1,0 +1,20 @@
+package com.vulpix.api.Dto.Publicacao.Insights;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.util.Map;
+
+@JsonIgnoreProperties(ignoreUnknown = true)
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class PublicacaoInsightDto {
+
+    private Map<String, ValueDto> metrics;
+}
+
