+++ src/main/java/com/vulpix/api/Dto/Publicacao/Insights/ValueDto.java
@@ -1,0 +1,17 @@
+package com.vulpix.api.Dto.Publicacao.Insights;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@JsonIgnoreProperties(ignoreUnknown = true)
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class ValueDto {
+
+    private int value;
+}
