--- src/main/java/com/vulpix/api/Dto/Publicacao/PublicacaoInsightDto.java
@@ -1,45 +1,0 @@
-package com.vulpix.api.Dto.Publicacao;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@JsonIgnoreProperties(ignoreUnknown = true)
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class PublicacaoInsightDto {
-
-    private List<InsightDataDto> data;
-}
-
-@JsonIgnoreProperties(ignoreUnknown = true)
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-class InsightDataDto {
-
-    private String name;
-    private String period;
-    private List<ValueDto> values;
-    private String title;
-    private String description;
-    private String id;
-}
-
-@JsonIgnoreProperties(ignoreUnknown = true)
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-class ValueDto {
-
-    private int value;
-}
